+++ Globals.h
@@ -5,9 +5,9 @@
 
 //if a reading has 3 components (like gyro or accel)
 struct three_axis {
-    int x;
-    int y;
-    int z;
+    float x;
+    float y;
+    float z;
 };
 
 //Fill this in once we know what the data will look like
@@ -17,9 +17,9 @@ struct datastruct {
     int data2 = 0;
 
     //Altimeter - MS5607 
-    int pressure = 0;
-    int temperature = 0;
-    int altitude = 0;
+    float pressure = 0;
+    float temperature = 0;
+    float altitude = 0;
 
     //Flight Profile - Adafruit IMU
     three_axis gyro =   (three_axis) {.x=0, .y=0, .z=0};
